refactor: split template page components
